requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
typing-extensions>=4.9.0
